requirements.txt
python-telegram-bot==22.0
aiosqlite==0.21.0
aiohttp==3.11.18 